requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pillow>=8.3.0

# Data handling
datasets>=1.18.0
opencv-python>=4.5.0

# Deep learning utilities  
segmentation-models-pytorch>=0.3.0
timm>=0.6.0

# Visualization and metrics
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0
wandb>=0.12.0

# Scientific computing
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.1.0
hydra-core>=1.1.0
tqdm>=4.62.0